CONSOLE: console DELAY now does event processing
